refactored stac validator
