+++ search/tests/validator/schemaValidator.spec.js
@@ -5,8 +5,6 @@ const {
   getSchema,
   getSchemaCollection,
   validateSchema,
-  createStacValidator,
-  validateStac
 } = require('../../lib/validator');
 
 describe('createSchemaValidator', () => {
@@ -199,51 +197,5 @@ describe('validateSchema', () => {
 
   it('should validate the test object against the schema for the given component', () => {
     expect(validateSchema('bbox', testBbox)).toEqual(true);
-  });
-});
-
-describe('createStacValidator', () => {
-  it('should exist', () => {
-    expect(createStacValidator).toBeDefined();
-  });
-
-  it('should take in a stacItem as a parameter', () => {
-    expect(() => createStacValidator()).toThrow();
-  });
-});
-
-describe('validateStac', () => {
-  const validStacFeature = {
-    stac_version: '0.8.0',
-    id: '1',
-    description: 'description',
-    links: []
-  };
-
-  const invalidStacFeature = {
-    stac_version: '0.8.0',
-    id: '1',
-    description: '',
-    links: []
-  };
-
-  it('should exist', () => {
-    expect(validateStac).toBeDefined();
-  });
-
-  it('should take in a stacItem as a parameter', () => {
-    expect(() => validateStac()).toThrow();
-  });
-
-  it('should take in a stacItem and return a boolean', () => {
-    expect(typeof validateStac(validStacFeature)).toEqual('boolean');
-  });
-
-  it('should check for required fields', () => {
-    expect(validateStac(invalidStacFeature)).toEqual(false);
-  });
-
-  it('should validate a stacItem with required fields', () => {
-    expect(validateStac(validStacFeature)).toEqual(true);
   });
 });
